Moved a bunch of database trait implementations out of nimiq-database and into their respective crates. This way nimiq-database doesn't depend on any other nimiq crate. Avoids circular dependencies.
