+++ database/src/traits.rs
@@ -4,21 +4,6 @@ use std::io;
 
 use lmdb_zero::traits::AsLmdbBytes;
 
-#[cfg(feature = "hash")]
-mod hash;
-
-#[cfg(feature = "block")]
-mod block;
-
-#[cfg(feature = "account")]
-mod accounts;
-
-#[cfg(feature = "keys")]
-mod keys;
-
-#[cfg(feature = "otp")]
-mod otp;
-
 pub trait IntoDatabaseValue {
     fn database_byte_size(&self) -> usize;
     fn copy_into_database(&self, bytes: &mut [u8]);
