+++ primitives/account/src/lib.rs
@@ -652,6 +652,26 @@ impl From<Vec<Receipt>> for Receipts {
     }
 }
 
+impl IntoDatabaseValue for Receipts {
+    fn database_byte_size(&self) -> usize {
+        self.serialized_size()
+    }
+
+    fn copy_into_database(&self, mut bytes: &mut [u8]) {
+        Serialize::serialize(&self, &mut bytes).unwrap();
+    }
+}
+
+impl FromDatabaseValue for Receipts {
+    fn copy_from_database(bytes: &[u8]) -> io::Result<Self>
+    where
+        Self: Sized,
+    {
+        let mut cursor = io::Cursor::new(bytes);
+        Ok(Deserialize::deserialize(&mut cursor)?)
+    }
+}
+
 #[derive(Clone, Eq, Debug, Serialize, Deserialize)]
 pub struct PrunedAccount {
     pub address: Address,
