+++ primitives/block/src/micro_block.rs
@@ -105,6 +105,26 @@ impl MicroBody {
     }
 }
 
+impl IntoDatabaseValue for MicroBlock {
+    fn database_byte_size(&self) -> usize {
+        self.serialized_size()
+    }
+
+    fn copy_into_database(&self, mut bytes: &mut [u8]) {
+        Serialize::serialize(&self, &mut bytes).unwrap();
+    }
+}
+
+impl FromDatabaseValue for MicroBlock {
+    fn copy_from_database(bytes: &[u8]) -> io::Result<Self>
+    where
+        Self: Sized,
+    {
+        let mut cursor = io::Cursor::new(bytes);
+        Ok(Deserialize::deserialize(&mut cursor)?)
+    }
+}
+
 impl Hash for MicroHeader {}
 
 impl fmt::Display for MicroHeader {
