+++ wallet/Cargo.toml
@@ -18,7 +18,7 @@ hex = "0.4"
 beserial = { path = "../beserial" }
 beserial_derive = { path = "../beserial/beserial_derive" }
 nimiq-account = { path = "../primitives/account" }
-nimiq-database = { path = "../database", features = ["keys", "otp"] }
+nimiq-database = { path = "../database" }
 nimiq-hash = { path = "../hash" }
 nimiq-key-derivation = { path = "../key-derivation" }
 nimiq-keys = { path = "../keys" }
